requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2; python_version >= "3.13"
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.25.1
fonttools==4.62.0
fsspec==2026.2.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.6.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.2.3
openai==2.26.0
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pypinyin==0.54.0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
ruff==0.15.5
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy>=2.0.36
starlette==0.52.1
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
zipp==3.23.0
scikit-learn>=1.4.0
psycopg[binary]>=3.2.3
